🐛 surface external action dispatch failures
