+++ apps/opencrane/src/app/external-action-executor.ts
@@ -1,24 +1,7 @@
 import type { JsonValue } from "@opencrane/util";
 import type { RuntimeExternalActionCandidate } from "@opencrane/contracts";
 import type { ExternalActionExecutor } from "@opencrane/backend/agents/runtime";
-import type { ObotCustodyPort } from "@opencrane/server/_infra/obot-custody";
-import type { SandboxJobExecutor } from "@opencrane/server/_infra/sandbox-execution";
-import type { MemoryGatewayClient } from "@opencrane/server/_infra/memory-gateway-client";
-
-/** Concrete transports the composition root injects into the external-action router. */
-export interface ExternalActionExecutorDependencies
-{
-	/** Silo owning the invocation, used as remote correlation context. */
-	readonly siloId: string;
-	/** Subject on whose behalf the action runs. */
-	readonly subjectId: string;
-	/** Obot credential-custody transport backing MCP tool calls (fail-closed until verified). */
-	readonly obotCustody: ObotCustodyPort;
-	/** Sandbox Job transport backing sandboxed tool calls (fail-closed until verified). */
-	readonly sandboxExecutor: SandboxJobExecutor;
-	/** Memory-gateway transport backing memory tool calls (fail-closed until verified). */
-	readonly memoryGateway: MemoryGatewayClient;
-}
+import type { ExternalActionExecutorDependencies } from "./external-action-executor.types.js";
 
 /** Typed failure raised for a candidate whose tool revision names no wired transport kind. */
 export class UnsupportedExternalActionError extends Error
